New test directory; messes up coverage
