+++ .circleci/config.yml
@@ -14,7 +14,7 @@ jobs:
 
       - run: docker build -t stevenmburns/tally_image .
 
-      - run: docker run --name test-container stevenmburns/tally_image bash -c "source /general/bin/activate && mkdir -p /test-reports && pytest --junitxml=test-reports/junit.xml -- /general/lib/python3.6/site-packages/tally/tally.py"
+      - run: docker run --name test-container stevenmburns/tally_image bash -c "source /general/bin/activate && mkdir -p /test-reports && pytest --junitxml=test-reports/junit.xml"
 
       - run: mkdir -p test-reports/pytest && docker cp test-container:/test-reports/junit.xml test-reports/pytest
 
@@ -27,7 +27,7 @@ jobs:
       - run: |
           ci_env=`curl -s https://codecov.io/env | bash`
           echo "ci_env: $ci_env"
-          docker run -e CODACY_PROJECT_TOKEN $ci_env --name coverage-container stevenmburns/tally_image bash -c "source /general/bin/activate && pytest --cov=tally -rs -- /general/lib/python3.6/site-packages/tally/tally.py && coverage html && coverage-badge -o htmlcov/coverage.svg && coverage xml && bash <(curl -s https://codecov.io/bash) && python-codacy-coverage -r coverage.xml"
+          docker run -e CODACY_PROJECT_TOKEN $ci_env --name coverage-container stevenmburns/tally_image bash -c "source /general/bin/activate && pytest --cov=tally -rs && coverage html && coverage-badge -o htmlcov/coverage.svg && coverage xml && bash <(curl -s https://codecov.io/bash) && python-codacy-coverage -r coverage.xml"
 
       - run: mkdir -p coverage-reports && docker cp coverage-container:/htmlcov coverage-reports/
